+++ package.json
@@ -20,7 +20,8 @@
         "axios": "^1.6.2",
         "cors": "^2.8.5",
         "dotenv": "^16.3.1",
-        "eslint": "^8.52.0"
+        "eslint": "^8.52.0",
+        "mongodb": "^6.3.0"
     },
     "devDependencies": {
         "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
