+++ packages/organizzen-react/src/MyApp.js
@@ -4,20 +4,22 @@ import NewEvent from './components/NewEvent'
 import NewTask from './components/NewTask'
 import './MyApp.css'
 
+const API_BASE_URL = 'https://organizzen.azurewebsites.net/'
 const App = () => {
     const [events, setEvents] = useState([])
 
+    // Replace the fetch URLs in MyApp.js
     useEffect(() => {
         // Fetch the list of events when the component mounts
-        fetch('http://localhost:8000/events')
+        fetch('${API_BASE_URL}/events')
             .then((response) => response.json())
-            .then((data) => setEvents(data.events_list))
+            .then((data) => setEvents(data)) // Update this line
             .catch((error) => console.log(error))
     }, []) // Empty dependency array ensures it runs only once on mount
 
     function postEvent(event) {
         // Add the new event to the backend
-        fetch('http://localhost:8000/events', {
+        fetch('${API_BASE_URL}/events', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
@@ -26,17 +28,16 @@ const App = () => {
         })
             .then(() => {
                 // After successfully adding the event, fetch the updated list
-                return fetch('http://localhost:8000/events')
+                return fetch('${API_BASE_URL}/events')
             })
             .then((response) => response.json())
-            .then((data) => setEvents(data.events_list))
+            .then((data) => setEvents(data)) // Update this line
             .catch((error) => console.log(error))
     }
 
     function postTask(eventId, task) {
         // Add the new task to the backend
-        fetch(`http://localhost:8000/events/${eventId}/tasks`, {
-            // <-- Note '/tasks' here
+        fetch(`${API_BASE_URL}/events/${eventId}/tasks`, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
@@ -45,14 +46,14 @@ const App = () => {
         })
             .then(() => {
                 // After successfully adding the task, fetch the updated list
-                return fetch(`http://localhost:8000/events/${eventId}/tasks`)
+                return fetch(`${API_BASE_URL}/events/${eventId}/tasks`)
             })
             .then((response) => response.json())
             .then((data) => {
                 // Update the tasks for the specific event in the state
                 setEvents((prevEvents) =>
                     prevEvents.map((event) =>
-                        event.id === eventId ? { ...event, tasks: data } : event
+                        event._id === eventId ? { ...event, tasks: data } : event
                     )
                 )
             })
@@ -62,7 +63,7 @@ const App = () => {
     return (
         <div>
             <HomePage />
-            <div className="newButtonContainer">
+            <div className="newButtons">
                 <div className="ButtonsContainer">
                     <NewEvent handleSubmit={postEvent} />
                     {/* Pass addTask function as a prop to NewTask component */}
@@ -72,6 +73,18 @@ const App = () => {
                 <div className="ButtonOtherDivider" />
                 <div className="TaskWidth" />
             </div>
+            {/* Render the list of events and tasks */}
+            {events.map((event) => (
+                <div key={event._id}>
+                    <h3>{event.name}</h3>
+                    <ul>
+                        {event.tasks.map((task) => (
+                            <li key={task._id}>{task.name}</li>
+                        ))}
+                    </ul>
+                </div>
+            ))}
+
         </div>
     )
 }
