+++ packages/organizzen-react/src/testMyApp.js
@@ -7,18 +7,17 @@ import './MyApp.css'
 const App = () => {
     const [events, setEvents] = useState([])
 
-    // Replace the fetch URLs in MyApp.js
     useEffect(() => {
         // Fetch the list of events when the component mounts
-        fetch('https://organizzen.azurewebsites.net/')
+        fetch('http://localhost:8000/events')
             .then((response) => response.json())
-            .then((data) => setEvents(data)) // Update this line
+            .then((data) => setEvents(data.events_list))
             .catch((error) => console.log(error))
     }, []) // Empty dependency array ensures it runs only once on mount
 
     function postEvent(event) {
         // Add the new event to the backend
-        fetch('https://organizzen.azurewebsites.net/', {
+        fetch('http://localhost:8000/events', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
@@ -27,16 +26,17 @@ const App = () => {
         })
             .then(() => {
                 // After successfully adding the event, fetch the updated list
-                return fetch('https://organizzen.azurewebsites.net/');
+                return fetch('http://localhost:8000/events')
             })
             .then((response) => response.json())
-            .then((data) => setEvents(data)) // Update this line
+            .then((data) => setEvents(data.events_list))
             .catch((error) => console.log(error))
     }
 
     function postTask(eventId, task) {
         // Add the new task to the backend
-        fetch(`https://organizzen.azurewebsites.net/`, {
+        fetch(`http://localhost:8000/events/${eventId}/tasks`, {
+            // <-- Note '/tasks' here
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
@@ -45,24 +45,24 @@ const App = () => {
         })
             .then(() => {
                 // After successfully adding the task, fetch the updated list
-                return fetch(`https://organizzen.azurewebsites.net/`);
+                return fetch(`http://localhost:8000/events/${eventId}/tasks`)
             })
             .then((response) => response.json())
             .then((data) => {
                 // Update the tasks for the specific event in the state
                 setEvents((prevEvents) =>
                     prevEvents.map((event) =>
-                        event._id === eventId ? { ...event, tasks: data } : event
+                        event.id === eventId ? { ...event, tasks: data } : event
                     )
                 )
             })
-            .catch((error) => console.log(error));
+            .catch((error) => console.log(error))
     }
 
     return (
         <div>
             <HomePage />
-            <div className="newButtons">
+            <div className="newButtonContainer">
                 <div className="ButtonsContainer">
                     <NewEvent handleSubmit={postEvent} />
                     {/* Pass addTask function as a prop to NewTask component */}
@@ -72,18 +72,6 @@ const App = () => {
                 <div className="ButtonOtherDivider" />
                 <div className="TaskWidth" />
             </div>
-            {/* Render the list of events and tasks */}
-            {events.map((event) => (
-                <div key={event._id}>
-                    <h3>{event.name}</h3>
-                    <ul>
-                        {event.tasks.map((task) => (
-                            <li key={task._id}>{task.name}</li>
-                        ))}
-                    </ul>
-                </div>
-            ))}
-
         </div>
     )
 }
